Programming Exercise/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.5
djangorestframework==3.15.2
kombu==5.4.2
prompt_toolkit==3.0.50
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
scikit-learn==1.3.0  
pandas==2.1.0      
django-cors-headers==4.0.0